rename: 삭제 - ROUTES
1. 폴더 '라우터' 삭제함
- SPA 전환에 따라 필요없어짐
2. var -> const 변경 작업
3. static 미들웨어 코드 위치를 라우팅 이후로 이동함
- 파일 처리보다 라우팅이 우선되어야 index.html과의 충돌이 해결됨
